requirements.txt
beautifulsoup4==4.9.3
lxml==4.6.1
requests==2.25.0
urllib3==1.26.2
pathvalidate==2.3.0
python-dotenv==0.15.0
tqdm==4.54.0